# application-api/src/main/resources/application.api.spec.json
{
  "swagger": "2.0",
  "info": {
    "title": "Application Server API",
    "version": "1.0",
    "description": "API для Application Server, который будет посередине между клиентом и узлами сети блокчейн"
  },
  "host": "example.com",
  "paths": {
    "/nonces": {
      "get": {
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "type": "object",
              "properties": {
                "begin_nonce": {
                  "type": "integer",
                  "format": "int64",
                  "description": "Начало диапазона nonce"
                },
                "end_nonce": {
                  "type": "integer",
                  "format": "int64",
                  "description": "Конец диапазона nonce"
                }
              },
              "required": [
                "end_nonce"
              ]
            },
            "examples": {
              "application/json": {
                "end_nonce": 10,
                "begin_nonce": 30
              }
            }
          }
        },
        "description": "Получение каждой нодой диапазона nonce для майнинга",
        "operationId": "getNonces",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "type": "object",
              "properties": {
                "node_id": {
                  "type": "integer",
                  "description": "Идентификатор узла"
                },
                "block_id": {
                  "type": "integer",
                  "format": "int64",
                  "description": "Идентификатор блока для добавления в блокчейн"
                }
              },
              "required": [
                "node_id",
                "block_id"
              ]
            },
            "x-examples": {
              "application/json": {
                "node_id": 1,
                "block_id": 101
              }
            }
          }
        ],
        "summary": "Получение nonce"
      }
    },
    "/files": {
      "post": {
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "type": "object",
              "properties": {}
            }
          }
        },
        "operationId": "uploadFile",
        "consumes": [
          "multipart/form-data"
        ],
        "description": "Отправка файла для загрузки",
        "parameters": [
          {
            "in": "header",
            "name": "User-Agent",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "file",
            "type": "file",
            "description": "Отправляемый файл"
          }
        ],
        "summary": "Загрузка файла"
      }
    },
    "/file/{file_hash}": {
      "get": {
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "type": "object",
              "properties": {}
            }
          }
        },
        "description": "Получение файла по его хэшу",
        "operationId": "getFile",
        "produces": [
          "application/octet-stream"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "file_hash",
            "type": "string",
            "required": true
          }
        ],
        "summary": "Получение файла"
      }
    }
  }
}